Выделит в презентере рендеринг компонентов
